pyproject.toml
[tool.poetry]
name = "parser-cian"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
cianparser = "^1.0.4"
pytest = "^8.2.1"
flake8 = "^7.0.0"
pytest-cov = "^5.0.0"
sqlalchemy = "^2.0.30"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
requests = "^2.32.3"
fake-headers = "^1.0.2"
tqdm = "^4.66.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
